// react-frontend/src/api/models/push-subscription-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spys API
 * REST API for Spys
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface PushSubscriptionRequest
 */
export interface PushSubscriptionRequest {
    /**
     * 
     * @type {string}
     * @memberof PushSubscriptionRequest
     */
    'endpoint': string;
    /**
     * 
     * @type {string}
     * @memberof PushSubscriptionRequest
     */
    'p256dh': string;
    /**
     * 
     * @type {string}
     * @memberof PushSubscriptionRequest
     */
    'auth': string;
    /**
     * 
     * @type {string}
     * @memberof PushSubscriptionRequest
     */
    'userAgent'?: string;
}